Renames the SlideList component to a more apropriate name
