+++ src/components/SlideListEditorComponent.js
